+++ demos/chinook/src/main/java/is/codion/demos/chinook/testing/ChinookLoadTest.java
@@ -33,13 +33,13 @@ import is.codion.demos.chinook.testing.scenarios.ViewGenre;
 import is.codion.demos.chinook.testing.scenarios.ViewInvoice;
 import is.codion.framework.db.EntityConnectionProvider;
 import is.codion.tools.loadtest.LoadTest;
-import is.codion.tools.loadtest.LoadTest.Scenario;
+import is.codion.tools.loadtest.Scenario;
 
 import java.util.Collection;
 import java.util.List;
 import java.util.function.Function;
 
-import static is.codion.tools.loadtest.LoadTest.Scenario.scenario;
+import static is.codion.tools.loadtest.Scenario.scenario;
 import static is.codion.tools.loadtest.model.LoadTestModel.loadTestModel;
 import static is.codion.tools.loadtest.ui.LoadTestPanel.loadTestPanel;