+++ demos/chinook/src/main/java/is/codion/demos/chinook/testing/scenarios/ViewCustomerReport.java
@@ -21,7 +21,7 @@ package is.codion.demos.chinook.testing.scenarios;
 import is.codion.demos.chinook.domain.api.Chinook.Customer;
 import is.codion.framework.db.EntityConnection;
 import is.codion.framework.db.EntityConnectionProvider;
-import is.codion.tools.loadtest.LoadTest.Scenario.Performer;
+import is.codion.tools.loadtest.Scenario.Performer;
 
 import java.util.HashMap;
 import java.util.List;