+++ demos/chinook/src/main/java/is/codion/demos/chinook/testing/scenarios/ViewGenre.java
@@ -23,7 +23,7 @@ import is.codion.demos.chinook.domain.api.Chinook.Track;
 import is.codion.framework.db.EntityConnection;
 import is.codion.framework.db.EntityConnectionProvider;
 import is.codion.framework.domain.entity.Entity;
-import is.codion.tools.loadtest.LoadTest.Scenario.Performer;
+import is.codion.tools.loadtest.Scenario.Performer;
 
 import java.util.ArrayList;
 import java.util.List;