+++ demos/employees/src/main/java/is/codion/demos/employees/testing/scenarios/LoginLogout.java
@@ -19,7 +19,7 @@
 package is.codion.demos.employees.testing.scenarios;
 
 import is.codion.demos.employees.model.EmployeesAppModel;
-import is.codion.tools.loadtest.LoadTest.Scenario.Performer;
+import is.codion.tools.loadtest.Scenario.Performer;
 
 import java.util.Random;