+++ tools/loadtest/core/src/main/java/is/codion/tools/loadtest/DefaultLoadTest.java
@@ -24,7 +24,7 @@ import is.codion.common.observer.Observer;
 import is.codion.common.state.State;
 import is.codion.common.user.User;
 import is.codion.common.value.Value;
-import is.codion.tools.loadtest.LoadTest.Scenario.Result;
+import is.codion.tools.loadtest.Scenario.Result;
 import is.codion.tools.loadtest.randomizer.ItemRandomizer;
 
 import org.slf4j.Logger;